adk-streaming/app/requirements.txt
google-adk>=0.6.0
google-genai>=0.8.0
google-cloud-dialogflow-cx>=1.40.0
python-dotenv>=1.0.0
